removed previous lobby state management + minor clean up of its residue
