+++ apps/nextjs/src/components/client/providers/index.ts
@@ -2,5 +2,4 @@ export { default as ThemeProvider } from "./ThemeProvider";
 export { default as DashboardTabProvider } from "./DashboardTabProvider";
 export { default as ControllerStateProvider } from "./ControllerStateProvider";
 export { default as FriendsChannelProvider } from "./FriendsChannelProvider";
-export { default as LobbyProvider } from "./LobbyProvider";
 export { default as MinaProvider } from "./MinaProvider";
